+++ CHANGES.md
@@ -6,7 +6,7 @@
   - `ImageData` has extra method to zoom image using information from a template image, `zoom_image_as_template`.
   - Error raised in `AcquisitionSensitivityModel.[un]normalise` methods applied to a read-only object.
   - `DataContainer.supports_array_view` to test for zero-copy compatibility.
-  - SIRF Python interface for STIR Poisson noise generation provided.
+  - SIRF interfaces (C++ and Python) for STIR Poisson noise generation utilities provided.
 
 ## v3.8.1